+++ src/app/features/admin/SanityReportDisplay.jsx
@@ -1,8 +1,28 @@
 import React, {useEffect} from "react";
 import {useGenerateSanityReportMutation} from "../../slices/api/adminApiSlice";
-import {getToastMessage} from "../../utils";
+import {downloadData, getToastMessage} from "../../utils";
 import {toast} from "react-toastify";
 import {SanityReport} from "./SanityReport";
+import {DownloadButton} from "../../components/controls/DownloadButton";
+
+
+const GenerateButton = (props) => {
+    let {isLoading, onClick} = props
+    let style = {
+        width: '8rem',
+        marginRight: '1rem',
+        cursor: isLoading ? 'wait' : 'pointer',
+    }
+    return (
+        <button
+            disabled={isLoading}
+            className="Small-button"
+            onClick={onClick}
+            style={style}>
+            Generate
+        </button>
+    )
+}
 
 export const SanityReportDisplay = () => {
 
@@ -12,6 +32,14 @@ export const SanityReportDisplay = () => {
         await generate()
         console.log('done generating report')
     }
+    const onDownloadReport = (report) => {
+        let timestamp = Math.floor(Date.now() / 1000);
+        let filename = `matchday_sanity_report_${timestamp}.json`
+        let reportData = JSON.stringify(report, null, 1)
+
+        console.log('Downloading sanity report...', filename)
+        downloadData(reportData, filename)
+    }
 
     const [generate, {
         data,
@@ -20,6 +48,7 @@ export const SanityReportDisplay = () => {
         isError,
         error
     }] = useGenerateSanityReportMutation()
+    let reportExists = isSuccess && data !== undefined && data !== null
 
     // toast messages
     useEffect(() => {
@@ -33,18 +62,15 @@ export const SanityReportDisplay = () => {
         <>
             <h1 style={{marginBottom: '1rem'}}>Sanity Report</h1>
             <p style={{color: '#aaa'}}>Click the button below to generate a System Sanity Check report.</p>
-            <button
-                disabled={isLoading}
-                className="Small-button"
-                onClick={onGenerateReport}
-                style={{width: '8rem', margin: '1rem 0'}}>
-                Generate
-            </button>
+            <div style={{display: 'flex', alignItems: 'center', marginTop: '1rem'}}>
+                <GenerateButton isLoading={isLoading} onClick={onGenerateReport}/>
+                {
+                    reportExists ? <DownloadButton onClick={() => onDownloadReport(data)}/> : null
+                }
+            </div>
             <div className="Sanity-report-display">
                 {
-                    isSuccess && data !== undefined ?
-                        <SanityReport report={data}/> :
-                        null
+                    reportExists ? <SanityReport report={data}/> : null
                 }
             </div>
         </>